requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Templates and static files
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client for source detection
httpx==0.25.2
requests==2.31.0

# Data processing
pandas==2.1.4
numpy==1.25.2

# Date and time handling
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For enhanced source detection
beautifulsoup4==4.12.2
lxml==4.9.3
yt-dlp>=2023.7.6  # For YouTube metadata extraction

# Optional: For async database operations
asyncpg==0.29.0

# Optional: For Redis caching (if needed later)
redis==5.0.1

# Optional: For background tasks
celery==5.3.4

# Optional: For monitoring and logging
structlog==23.2.0
